+++ lib/game/components/components.dart
@@ -1,1 +1,2 @@
+export 'pasture_field.dart';
 export 'unicorn_counter.dart';
