+++ lib/game/spawners/food_spawner.dart
@@ -4,11 +4,12 @@ import 'package:flame/components.dart';
 import 'package:flame_bloc/flame_bloc.dart';
 import 'package:ranch_components/ranch_components.dart';
 import 'package:ranch_flame/ranch_flame.dart';
+import 'package:very_good_ranch/game/components/components.dart';
 import 'package:very_good_ranch/game/entities/food/food.dart';
 import 'package:very_good_ranch/game/game.dart';
 
 class FoodSpawner extends TimerComponent
-    with ParentIsA<PastureArea>, FlameBlocListenable<GameBloc, GameState> {
+    with ParentIsA<PastureField>, FlameBlocListenable<GameBloc, GameState> {
   FoodSpawner({
     required this.seed,
     double spawnThreshold = 60.0,
