+++ test/game/spawners/unicorn_spawner_test.dart
@@ -5,9 +5,9 @@ import 'dart:math';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockingjay/mockingjay.dart';
+import 'package:very_good_ranch/game/components/components.dart';
 import 'package:very_good_ranch/game/entities/entities.dart';
 import 'package:very_good_ranch/game/spawners/spawners.dart';
-import 'package:very_good_ranch/game/very_good_ranch_game.dart';
 
 import '../../helpers/helpers.dart';
 
@@ -29,7 +29,7 @@ void main() {
       setUp: (game, tester) async {
         when(() => seed.nextDouble()).thenReturn(1);
         await game.add(
-          PastureArea(
+          PastureField(
             children: [
               UnicornSpawner(seed: seed),
             ],
@@ -41,7 +41,7 @@ void main() {
         await game.ready();
       },
       verify: (game, tester) async {
-        final pastureArea = game.children.whereType<PastureArea>().first;
+        final pastureArea = game.children.whereType<PastureField>().first;
         final unicornComponents = pastureArea.children.whereType<Unicorn>();
         expect(unicornComponents.length, 1);
         expect(unicornComponents.first.position, pastureArea.size);
